+++ lib/src/routers/router_delegate.dart
@@ -17,12 +17,17 @@ class QRouterDelegate extends RouterDelegate<String> with ChangeNotifier {
     this.withWebBar = false,
     this.alwaysAddInitPath = false,
     List<NavigatorObserver>? observers,
+    this.restorationScopeId,
   }) : key = navKey ?? GlobalKey<NavigatorState>() {
     _createController();
     if (observers != null) {
       this.observers.addAll(observers);
     }
   }
+
+  /// Restoration ID to save and restore the state of the navigator, including
+  /// its history.
+  final String? restorationScopeId;
 
   /// Set this to true if you want always the initial router to be added on the stack
   /// Example.
@@ -112,15 +117,22 @@ class QRouterDelegate extends RouterDelegate<String> with ChangeNotifier {
     return;
   }
 
-  Navigator get navigator => Navigator(
-        key: key,
-        pages: _controller.pages,
-        observers: observers,
-        onPopPage: (route, result) {
-          _controller.removeLast();
-          return false;
-        },
-      );
+  Navigator get navigator {
+    var scopId = restorationScopeId;
+    if (scopId == null && QR.settings.autoRestoration) {
+      scopId = 'router:${_controller.key.name}';
+    }
+    return Navigator(
+      key: key,
+      pages: _controller.pages,
+      observers: observers,
+      restorationScopeId: restorationScopeId,
+      onPopPage: (route, result) {
+        _controller.removeLast();
+        return false;
+      },
+    );
+  }
 
   Widget _buildNavigator() {
     if (!withWebBar || !BrowserAddressBar.isNeeded) {