+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix #111
 - Fix #112
+- Add restorationScopeId for the routers and auto restoration
 
 ## 1.8.1  25.02.2023